feat: MVC Model 적용 및 Error 파일 생성
