+++ Calculator/Calculator/Model/Error.swift
@@ -1,0 +1,12 @@
+//
+//  Error.swift
+//  Calculator
+//
+//  Created by 정연호 on 2022/09/07.
+//
+
+import Foundation
+
+enum CalculatorError : Error {
+    case noValue
+}
