+++ Calculator/CalculatorTests/CalculatorTests.swift
@@ -7,46 +7,6 @@
 
 import XCTest
 
-protocol CalculateItem {
-    
-}
-
-enum CalculatorError : Error {
-    case noValue
-}
-
-struct CalculatorItemQueue<T>: CalculateItem {
-    var list: [T] = []
-    
-    mutating func enqueue(_ item: T) {
-        list.append(item)
-    }
-    
-    mutating func dequeue() -> T {
-        return list.removeFirst()
-    }
-    
-    mutating func clear() {
-        list.removeAll()
-    }
-    
-    func peak() throws -> T {
-        if let item = list.first {
-            return item
-        } else {
-            throw CalculatorError.noValue
-        }
-    }
-    
-    func count() -> Int {
-        return list.count
-    }
-    
-    func isEmpty() -> Bool {
-        return list.isEmpty
-    }
-}
-
 class CalculatorTests: XCTestCase {
 
     override func setUpWithError() throws {
